+++ packages/web-components/src/components/footer/footer-composite.ts
@@ -92,10 +92,22 @@ class DDSFooterComposite extends ModalRenderMixin(HybridRenderMixin(HostListener
   _setLanguage?: (language: string) => void;
 
   /**
+   * The aria-label to use for the locale-button
+   */
+  @property()
+  buttonLabel?: string;
+
+  /**
    * The g11n collator to use for sorting contry names.
    */
   @property({ attribute: false })
   collatorCountryName = new Intl.Collator();
+
+  /**
+   * `true` to omit the locale switcher button.
+   */
+  @property({ attribute: 'disable-locale-button' })
+  disableLocaleButton = false;
 
   /**
    * The language used for query.
@@ -187,7 +199,15 @@ class DDSFooterComposite extends ModalRenderMixin(HybridRenderMixin(HostListener
   }
 
   renderLightDOM() {
-    const { langDisplay, size, links, legalLinks, _handleClickLocaleButton: handleClickLocaleButton } = this;
+    const {
+      buttonLabel,
+      disableLocaleButton,
+      langDisplay,
+      size,
+      links,
+      legalLinks,
+      _handleClickLocaleButton: handleClickLocaleButton,
+    } = this;
     return html`
       <dds-footer size="${ifNonNull(size)}">
         <dds-footer-logo></dds-footer-logo>
@@ -204,9 +224,11 @@ class DDSFooterComposite extends ModalRenderMixin(HybridRenderMixin(HostListener
             `
           )}
         </dds-footer-nav>
-        ${size !== FOOTER_SIZE.MICRO
+        ${!disableLocaleButton && size !== FOOTER_SIZE.MICRO
           ? html`
-              <dds-locale-button @click="${handleClickLocaleButton}">${langDisplay}</dds-locale-button>
+              <dds-locale-button buttonLabel="${ifNonNull(buttonLabel)}" @click="${handleClickLocaleButton}"
+                >${langDisplay}</dds-locale-button
+              >
             `
           : html``}
         <dds-legal-nav size="${ifNonNull(size)}">
@@ -218,7 +240,11 @@ class DDSFooterComposite extends ModalRenderMixin(HybridRenderMixin(HostListener
           <dds-legal-nav-cookie-preferences-placeholder></dds-legal-nav-cookie-preferences-placeholder>
           ${size === FOOTER_SIZE.MICRO
             ? html`
-                <dds-locale-button size="${size}" slot="locale" @click="${handleClickLocaleButton}"
+                <dds-locale-button
+                  buttonLabel="${ifNonNull(buttonLabel)}"
+                  size="${size}"
+                  slot="locale"
+                  @click="${handleClickLocaleButton}"
                   >${langDisplay}</dds-locale-button
                 >
               `