+++ packages/web-components/src/components/locale-modal/locale-modal.ts
@@ -45,8 +45,9 @@ class DDSLocaleModal extends DDSExpressiveModal {
   /**
    * Handles `click` event on the back button.
    */
-  private _handleClickBackButton() {
+  private _handleClickBackButton(e) {
     this._currentRegion = undefined;
+    e.preventDefault();
   }
 
   /**
@@ -55,7 +56,10 @@ class DDSLocaleModal extends DDSExpressiveModal {
    * @param event The event.
    */
   private _handleClickRegionSelector(event: MouseEvent) {
-    this._currentRegion = (event.target as DDSRegionItem).name;
+    const { invalid, name } = event.target as DDSRegionItem;
+    if (!invalid) {
+      this._currentRegion = name;
+    }
   }
 
   @HostListener('eventClose')
@@ -92,7 +96,7 @@ class DDSLocaleModal extends DDSExpressiveModal {
   private _renderLocaleSelectorHeading() {
     const { headerTitle, _currentRegion: currentRegion, _handleClickBackButton: handleClickBackButton } = this;
     return html`
-      <dds-link-with-icon icon-placement="${ICON_PLACEMENT.LEFT}" href="#" ?disabled="false" @click="${handleClickBackButton}">
+      <dds-link-with-icon icon-placement="${ICON_PLACEMENT.LEFT}" href="#" @click="${handleClickBackButton}">
         ${headerTitle}${ArrowLeft20({ slot: 'icon', class: `${prefix}--locale-modal__label-arrow` })}
       </dds-link-with-icon>
       <p class="bx--modal-header__heading bx--type-beta" tabindex="0">${currentRegion}</p>