+++ packages/services-store/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.14.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/master/packages/services-store/compare/@carbon/ibmdotcom-services-store@1.14.0-rc.0...@carbon/ibmdotcom-services-store@1.14.0) (2020-12-15)
+
+**Note:** Version bump only for package @carbon/ibmdotcom-services-store
+
+# [1.14.0-rc.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/master/packages/services-store/compare/@carbon/ibmdotcom-services-store@1.13.0...@carbon/ibmdotcom-services-store@1.14.0-rc.0) (2020-12-04)
+
+### Features
+
+- **masthead:** alternative logo in web components ([#4581](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/master/packages/services-store/issues/4581)) ([e958b87](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/master/packages/services-store/commit/e958b87)), closes [#4257](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/master/packages/services-store/issues/4257)
+- **masthead:** masthead L1 web components ([#4429](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/master/packages/services-store/issues/4429)) ([e0bff75](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/master/packages/services-store/commit/e0bff75)), closes [#3806](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/master/packages/services-store/issues/3806)
+
 # [1.13.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/master/packages/services-store/compare/@carbon/ibmdotcom-services-store@1.13.0-rc.0...@carbon/ibmdotcom-services-store@1.13.0) (2020-11-25)
 
 **Note:** Version bump only for package @carbon/ibmdotcom-services-store